customer-account-ui-extensions: add metadata to the standard api
