+++ packages/customer-account-ui-extensions/src/extension-points/index.ts
@@ -1,11 +1,12 @@
 export type {ExtensionPoints, ExtensionPoint} from './extension-points';
 export type {
-  StandardApi,
   Language,
   Localization,
   I18nTranslate,
   I18n,
-} from './standard-api';
+} from './standard-api/localization';
+export type {StandardApi} from './standard-api';
+
 export type {
   ApiForExtension,
   ArgumentsForExtension,
